Move assets to their own folder to avoid import cycle
Deduplicates the toml import by assets and locals
