--- i18n/assets.go
@@ -1,8 +1,0 @@
-package i18n
-
-import (
-	"embed"
-)
-
-//go:embed *.toml
-var LocaleFS embed.FS
